requirements.txt
requests
pillow
reportlab
google-play-scraper